refactor: 전역 설정 컴포넌트 분리 #39
