+++ app/_layout.tsx
@@ -15,56 +15,10 @@ import FontAwesome6 from "@expo/vector-icons/FontAwesome6";
 import { useAuthStore } from "@/domain/auth/stores/useAuthStore";
 import path from "@/common/constants/path";
 import AuthEventListener from "./../src/domain/auth/events/authEventListener";
+import AuthStateListener from "@/common/global/authStateListener";
+import LoadingScreen from "@/common/global/LoadingScreen";
 
 SplashScreen.preventAutoHideAsync().catch(() => {});
-
-// 인증 상태에 따라 리디렉션하는 컴포넌트
-function AuthStateListener({ children }: { children: ReactNode }) {
-  const router = useRouter();
-  const segments = useSegments();
-  const { initAuth, isAuthenticated, isLoading } = useAuthStore();
-  const initialCheckRef = useRef(false);
-
-  // 초기 인증 체크를 위한 useEffect
-  useEffect(() => {
-    if (isLoading || initialCheckRef.current) return;
-
-    const checkAuth = async () => {
-      await initAuth();
-      initialCheckRef.current = true;
-    };
-
-    checkAuth();
-  }, [isLoading, initAuth]);
-
-  // 통합된 라우팅 처리
-  useEffect(() => {
-    if (isLoading || !initialCheckRef.current) return;
-
-    const isProtectedRoute = segments[0] === ("(protected)" as string);
-    const isPublicRoute = segments[0] === ("(public)" as string);
-    const isRootRoute =
-      segments.length === (0 as number) ||
-      (segments.length === 1 && segments[0] === ("index" as string));
-
-    // 인증되지 않은 상태에서 protected 라우트 접근
-    if (!isAuthenticated && isProtectedRoute) {
-      console.log("Redirecting to welcome page - not authenticated");
-      router.replace(path.showWelcome());
-    }
-    // 인증된 상태에서 public 또는 root 라우트 접근
-    else if (isAuthenticated && (isPublicRoute || isRootRoute)) {
-      console.log("Redirecting to room list - authenticated");
-      router.replace(path.showRoomList());
-    }
-  }, [isAuthenticated, isLoading, segments, router]);
-  // 최초 로딩 중일 때만 로딩 화면 표시
-  if (isLoading && !initialCheckRef.current) {
-    return <LoadingScreen />;
-  }
-
-  return <>{children}</>;
-}
 
 export default function RootLayout() {
   const [fontsLoaded, fontsError] = useFonts({
@@ -114,13 +68,5 @@ export default function RootLayout() {
         </CustomQueryClientProvider>
       </ErrorBoundary>
     </PaperProvider>
-  );
-}
-
-function LoadingScreen() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <ActivityIndicator size="large" color="#000" />
-    </View>
   );
 }
