+++ src/common/global/LoadingScreen.tsx
@@ -1,0 +1,9 @@
+import { View, ActivityIndicator } from "react-native";
+
+export default function LoadingScreen() {
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <ActivityIndicator size="large" color="#000" />
+    </View>
+  );
+}
